consolidate widgets creation into one file
